# public/swagger.json
{
	"components": {
...
		"schemas": {
			"Types.ObjectId": {
				"type": "string"
			},
...
			"ISession": {
				"properties": {
					"_id": {
						"anyOf": [
							{
								"$ref": "#/components/schemas/Types.ObjectId"
							},
							{
								"type": "string"
							}
						]
					},
					"sessionType": {
						"type": "string",
						"enum": [
							"group",
							"individual"
						]
					},
					"serviceProviderID": {
						"anyOf": [
							{
								"type": "string"
							},
							{
								"$ref": "#/components/schemas/Types.ObjectId"
							},
							{
								"$ref": "#/components/schemas/IServiceProvider"
							}
						]
					},
					"clientsID": {
						"anyOf": [
							{
								"items": {
									"type": "string"
								},
								"type": "array"
							},
							{
								"items": {
									"$ref": "#/components/schemas/Types.ObjectId"
								},
								"type": "array"
							},
							{
								"items": {
									"$ref": "#/components/schemas/IClient"
								},
								"type": "array"
							}
						]
					},
					"name": {
						"type": "string"
					},
					"details": {
						"type": "string"
					},
					"startDate": {
						"type": "string",
						"format": "date-time"
					},
					"duration": {
						"type": "number",
						"format": "double"
					},
					"serviceType": {
						"type": "string",
						"enum": [
							"online",
							"home",
							"office"
						]
					},
					"location": {
						"additionalProperties": false,
						"type": "object"
					},
					"attachments": {
						"items": {
							"properties": {
								"type": {
									"type": "string"
								},
								"name": {
									"type": "string"
								},
								"url": {
									"type": "string"
								}
							},
							"required": [
								"type",
								"name",
								"url"
							],
							"type": "object"
						},
						"type": "array"
					},
					"requirements": {
						"type": "string"
					},
					"ratings": {
						"items": {
							"properties": {
								"ratingDate": {
									"type": "string",
									"format": "date-time"
								},
								"ratingValue": {
									"type": "string"
								},
								"raterUID": {
									"anyOf": [
										{
											"type": "string"
										},
										{
											"$ref": "#/components/schemas/Types.ObjectId"
										},
										{
											"$ref": "#/components/schemas/IUser"
										}
									]
								}
							},
							"required": [
								"ratingDate",
								"ratingValue",
								"raterUID"
							],
							"type": "object"
						},
						"type": "array"
					},
					"reviews": {
						"items": {
							"properties": {
								"reviewDate": {
									"type": "string",
									"format": "date-time"
								},
								"reviewDetails": {
									"type": "string"
								},
								"reviewerUID": {
									"anyOf": [
										{
											"type": "string"
										},
										{
											"$ref": "#/components/schemas/Types.ObjectId"
										},
										{
											"$ref": "#/components/schemas/IUser"
										}
									]
								}
							},
							"required": [
								"reviewDate",
								"reviewDetails",
								"reviewerUID"
							],
							"type": "object"
						},
						"type": "array"
					},
					"sessionFee": {
						"type": "number",
						"format": "double"
					},
					"payments": {
						"items": {
							"properties": {
								"amount": {
									"type": "number",
									"format": "double"
								},
								"payerID": {
									"anyOf": [
										{
											"type": "string"
										},
										{
											"$ref": "#/components/schemas/Types.ObjectId"
										},
										{
											"$ref": "#/components/schemas/IUser"
										}
									]
								},
								"paymentMethod": {
									"type": "string"
								},
								"discount": {
									"type": "number",
									"format": "double"
								}
							},
							"required": [
								"amount",
								"payerID",
								"paymentMethod",
								"discount"
							],
							"type": "object"
						},
						"type": "array"
					},
					"status": {
						"type": "string",
						"enum": [
							"initiated",
							"agreed",
							"canceled",
							"finished"
						]
					},
					"doctorReferral": {
						"type": "string"
					},
					"url": {
						"type": "string"
					}
				},
				"required": [
					"sessionType",
					"name",
					"serviceType",
					"location",
					"requirements",
					"sessionFee",
					"status"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"IAgreement": {
				"properties": {
					"_id": {
						"anyOf": [
							{
								"type": "string"
							},
							{
								"$ref": "#/components/schemas/Types.ObjectId"
							}
						]
					},
					"name": {
						"type": "string"
					},
					"parties": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"startDate": {
						"type": "string",
						"format": "date-time"
					},
					"endDate": {
						"type": "string",
						"format": "date-time"
					},
					"details": {
						"type": "string"
					},
					"attachments": {
						"items": {
							"properties": {
								"disType": {
									"type": "string"
								},
								"name": {
									"type": "string"
								},
								"url": {
									"type": "string"
								}
							},
							"required": [
								"disType",
								"name",
								"url"
							],
							"type": "object"
						},
						"type": "array"
					},
					"reminder": {
						"type": "boolean"
					},
					"url": {
						"type": "string"
					}
				},
				"required": [
					"name",
					"parties",
					"startDate",
					"endDate",
					"details"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"Partial_IAgreement_": {
				"properties": {
					"_id": {
						"anyOf": [
							{
								"type": "string"
							},
							{
								"$ref": "#/components/schemas/Types.ObjectId"
							}
						]
					},
					"name": {
						"type": "string"
					},
					"parties": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"startDate": {
						"type": "string",
						"format": "date-time"
					},
					"endDate": {
						"type": "string",
						"format": "date-time"
					},
					"details": {
						"type": "string"
					},
					"attachments": {
						"items": {
							"properties": {
								"disType": {
									"type": "string"
								},
								"name": {
									"type": "string"
								},
								"url": {
									"type": "string"
								}
							},
							"required": [
								"disType",
								"name",
								"url"
							],
							"type": "object"
						},
						"type": "array"
					},
					"reminder": {
						"type": "boolean"
					},
					"url": {
						"type": "string"
					}
				},
				"type": "object",
				"description": "Make all properties in T optional"
			},
...
			"FlattenMaps_T_": {
				"properties": {},
				"type": "object"
			},
			"Document_any.any.any_": {
				"$ref": "#/components/schemas/FlattenMaps_T_",
				"description": "Generic types for Document:\n*  T - the type of _id\n*  TQueryHelpers - Object with any helpers that should be mixed into the Query type\n*  DocType - the type of the actual Document created"
			},
			"Required_any_": {
				"properties": {},
				"type": "object",
				"description": "Make all properties in T required"
			},
...
			"IAlarm": {
				"properties": {
					"_id": {
						"anyOf": [
							{
								"type": "string"
							},
							{
								"$ref": "#/components/schemas/Types.ObjectId"
							}
						]
					},
					"name": {
						"type": "string"
					},
					"referenceType": {
						"type": "string",
						"enum": [
							"Agreement",
							"Session",
							"Dispute",
							"ServiceProvider"
						]
					},
					"referenceID": {
						"anyOf": [
							{
								"type": "string"
							},
							{
								"$ref": "#/components/schemas/IAgreement"
							},
							{
								"$ref": "#/components/schemas/ISession"
							},
							{
								"$ref": "#/components/schemas/IDispute"
							},
							{
								"$ref": "#/components/schemas/IServiceProvider"
							}
						]
					},
					"frequencyUnit": {
						"type": "string",
						"enum": [
							"Days",
							"Hours"
						]
					},
					"frequency": {
						"type": "number",
						"format": "double"
					},
					"active": {
						"type": "boolean"
					},
					"startDate": {
						"type": "string",
						"format": "date-time"
					},
					"endDate": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"name",
					"referenceType",
					"referenceID",
					"frequencyUnit",
					"frequency",
					"active",
					"startDate",
					"endDate"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IAnnouncement": {
				"properties": {
					"_id": {
						"anyOf": [
							{
								"$ref": "#/components/schemas/Types.ObjectId"
							},
							{
								"type": "string"
							}
						]
					},
					"referenceType": {
						"type": "string",
						"enum": [
							"Session",
							"Advertisement"
						]
					},
					"referenceID": {
						"anyOf": [
							{
								"type": "string"
							},
							{
								"$ref": "#/components/schemas/ISession"
							},
							{
								"$ref": "#/components/schemas/INotification"
							}
						]
					},
					"status": {
						"type": "string",
						"enum": [
							"draft",
							"published"
						]
					},
					"topic": {
						"type": "string"
					},
					"details": {
						"type": "string"
					},
					"sentDate": {
						"type": "string",
						"format": "date-time"
					},
					"attachments": {
						"items": {
							"properties": {
								"attachmentType": {
									"type": "string"
								},
								"attachmentName": {
									"type": "string"
								},
								"attachmentUrl": {
									"type": "string"
								}
							},
							"required": [
								"attachmentType",
								"attachmentName",
								"attachmentUrl"
							],
							"type": "object"
						},
						"type": "array"
					},
					"receiversUIDs": {
						"anyOf": [
							{
								"items": {
									"type": "string"
								},
								"type": "array"
							},
							{
								"items": {
									"$ref": "#/components/schemas/Types.ObjectId"
								},
								"type": "array"
							},
							{
								"items": {
									"$ref": "#/components/schemas/IUser"
								},
								"type": "array"
							}
						]
					},
					"url": {
						"type": "string"
					}
				},
				"required": [
					"referenceType",
					"status",
					"topic",
					"details",
					"attachments"
				],
				"type": "object",
				"additionalProperties": false
			},
			"INotification": {
				"properties": {
					"_id": {
						"anyOf": [
							{
								"$ref": "#/components/schemas/Types.ObjectId"
							},
							{
								"type": "string"
							}
						]
					},
					"referenceType": {
						"type": "string",
						"enum": [
							"Session",
							"Agreement",
							"Dispute",
							"Alarm",
							"Announcement",
							"User",
							"Communication"
						]
					},
					"referenceID": {
						"anyOf": [
							{
								"type": "string"
							},
							{
								"$ref": "#/components/schemas/ISession"
							},
							{
								"$ref": "#/components/schemas/IAgreement"
							},
							{
								"$ref": "#/components/schemas/IDispute"
							},
							{
								"$ref": "#/components/schemas/IAlarm"
							},
							{
								"$ref": "#/components/schemas/IAnnouncement"
							},
							{
								"$ref": "#/components/schemas/IUser"
							},
							{
								"$ref": "#/components/schemas/ICommunication"
							}
						]
					},
					"status": {
						"type": "string",
						"enum": [
							"sent",
							"delivered",
							"opened"
						]
					},
					"title": {
						"type": "string"
					},
					"details": {
						"type": "string"
					},
					"sentDate": {
						"type": "string",
						"format": "date-time"
					},
					"receivedDate": {
						"type": "string",
						"format": "date-time"
					},
					"openDate": {
						"type": "string",
						"format": "date-time"
					},
					"receiverUID": {
						"anyOf": [
							{
								"type": "string"
							},
							{
								"$ref": "#/components/schemas/Types.ObjectId"
							},
							{
								"$ref": "#/components/schemas/IUser"
							}
						]
					}
				},
				"required": [
					"referenceType",
					"referenceID",
					"status",
					"title",
					"details",
					"sentDate",
					"receiverUID"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Partial_IAnnouncement_": {
				"properties": {
					"_id": {
						"anyOf": [
							{
								"$ref": "#/components/schemas/Types.ObjectId"
							},
							{
								"type": "string"
							}
						]
					},
					"referenceType": {
						"type": "string",
						"enum": [
							"Session",
							"Advertisement"
						]
					},
					"referenceID": {
						"anyOf": [
							{
								"type": "string"
							},
							{
								"$ref": "#/components/schemas/ISession"
							},
							{
								"$ref": "#/components/schemas/INotification"
							}
						]
					},
					"status": {
						"type": "string",
						"enum": [
							"draft",
							"published"
						]
					},
					"topic": {
						"type": "string"
					},
					"details": {
						"type": "string"
					},
					"sentDate": {
						"type": "string",
						"format": "date-time"
					},
					"attachments": {
						"items": {
							"properties": {
								"attachmentType": {
									"type": "string"
								},
								"attachmentName": {
									"type": "string"
								},
								"attachmentUrl": {
									"type": "string"
								}
							},
							"required": [
								"attachmentType",
								"attachmentName",
								"attachmentUrl"
							],
							"type": "object"
						},
						"type": "array"
					},
					"receiversUIDs": {
						"anyOf": [
							{
								"items": {
									"type": "string"
								},
								"type": "array"
							},
							{
								"items": {
									"$ref": "#/components/schemas/Types.ObjectId"
								},
								"type": "array"
							},
							{
								"items": {
									"$ref": "#/components/schemas/IUser"
								},
								"type": "array"
							}
						]
					},
					"url": {
						"type": "string"
					}
				},
				"type": "object",
				"description": "Make all properties in T optional"
			},
			"Partial_IAlarm_": {
				"properties": {
					"_id": {
						"anyOf": [
							{
								"type": "string"
							},
							{
								"$ref": "#/components/schemas/Types.ObjectId"
							}
						]
					},
					"name": {
						"type": "string"
					},
					"referenceType": {
						"type": "string",
						"enum": [
							"Agreement",
							"Session",
							"Dispute",
							"ServiceProvider"
						]
					},
					"referenceID": {
						"anyOf": [
							{
								"type": "string"
							},
							{
								"$ref": "#/components/schemas/IAgreement"
							},
							{
								"$ref": "#/components/schemas/ISession"
							},
							{
								"$ref": "#/components/schemas/IDispute"
							},
							{
								"$ref": "#/components/schemas/IServiceProvider"
							}
						]
					},
					"frequencyUnit": {
						"type": "string",
						"enum": [
							"Days",
							"Hours"
						]
					},
					"frequency": {
						"type": "number",
						"format": "double"
					},
					"active": {
						"type": "boolean"
					},
					"startDate": {
						"type": "string",
						"format": "date-time"
					},
					"endDate": {
						"type": "string",
						"format": "date-time"
					}
				},
				"type": "object",
				"description": "Make all properties in T optional"
			},
...
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "",
		"contact": {}
	},
	"openapi": "3.0.0",
	"paths": {
...
		"/agreements/{agreementId}": {
			"get": {
				"operationId": "GetAgreement",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/IAgreement"
										}
									],
									"nullable": true
								}
							}
						}
					},
					"404": {
						"description": "The requested agreement is not found"
					}
				},
				"description": "Get a agreement details",
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "agreementId",
						"required": true,
						"schema": {
							"type": "string"
						},
						"example": "630ded831cb815c4d98d3df6"
					}
				]
			},
			"delete": {
				"operationId": "DeleteAgreement",
				"responses": {
					"200": {
						"description": "Deleted",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/IAgreement"
										}
									],
									"nullable": true
								}
							}
						}
					},
					"404": {
						"description": "The requested agreement is not found"
					}
				},
				"description": "Delete a agreement",
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "agreementId",
						"required": true,
						"schema": {
							"type": "string"
						},
						"example": "630ded831cb815c4d98d3df6"
					}
				]
			}
		},
		"/agreements/create": {
			"post": {
				"operationId": "CreateAgreement",
				"responses": {
					"200": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IAgreement"
								},
								"examples": {
									"Example 1": {
										"value": {
											"name": "Ahmad",
											"parties": [
												"63124dffdcf1e4974079a441"
											],
											"startDate": "2022-09-10T00:00:00.000Z",
											"endDate": "2022-09-10T00:00:00.000Z",
											"details": "first Agreement",
											"attachments": [],
											"reminder": true
										}
									}
								}
							}
						}
					},
					"422": {
						"description": "Validation Failed"
					}
				},
				"description": "Create a agreement",
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IAgreement"
							}
						}
					}
				}
			}
		},
		"/agreements/update/{agreementId}": {
			"put": {
				"operationId": "UpdateAgreement",
				"responses": {
					"200": {
						"description": "Updated",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/IAgreement"
										}
									],
									"nullable": true
								}
							}
						}
					},
					"422": {
						"description": "Validation Failed"
					}
				},
				"description": "Update a agreement",
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "agreementId",
						"required": true,
						"schema": {
							"type": "string"
						},
						"example": "630ded831cb815c4d98d3df6"
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Partial_IAgreement_"
							},
							"example": "630ded831cb815c4d98d3df6"
						}
					}
				}
			}
		},
...
		"/announcements/{announcementId}": {
			"get": {
				"operationId": "GetAnnouncement",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/IAnnouncement"
										}
									],
									"nullable": true
								}
							}
						}
					},
					"404": {
						"description": "the requested Announcement in not found"
					}
				},
				"description": "Get a Announcement details",
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "announcementId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"delete": {
				"operationId": "DeleteAnnouncement",
				"responses": {
					"200": {
						"description": "Deleted",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/Document_any.any.any_"
										},
										{
											"$ref": "#/components/schemas/IAnnouncement"
										},
										{
											"$ref": "#/components/schemas/Required_any_"
										}
									]
								}
							}
						}
					},
					"404": {
						"description": "the requested announcement in not found"
					}
				},
				"description": "Delete a announcement",
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "announcementId",
						"required": true,
						"schema": {
							"type": "string"
						},
						"example": "631330dbb5a0b3d3c3d6b0e7"
					}
				]
			}
		},
		"/announcements/create": {
			"post": {
				"operationId": "CreateAnnouncement",
				"responses": {
					"200": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IAnnouncement"
								},
								"examples": {
									"Example 1": {
										"value": {
											"referenceType": "Session",
											"referenceID": "63131c361ea32f520cb28bd6",
											"status": "draft",
											"topic": "new topic",
											"details": "The announcement contain info ",
											"sentDate": "2022-09-10T00:00:00.000Z",
											"attachments": [],
											"receiversUIDs": [
												"63124dffdcf1e4974079a441",
												"63130ba28d796320172e58d6"
											]
										}
									}
								}
							}
						}
					},
					"422": {
						"description": "Validation Failed"
					}
				},
				"description": "Create a Announcement",
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IAnnouncement"
							}
						}
					}
				}
			}
		},
		"/announcements/update/{announcementId}": {
			"put": {
				"operationId": "UpdateAnnouncement",
				"responses": {
					"200": {
						"description": "updated",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/IAnnouncement"
										}
									],
									"nullable": true
								}
							}
						}
					},
					"422": {
						"description": "Validation Failed"
					}
				},
				"description": "Update a announcement",
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "announcementId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Partial_IAnnouncement_"
							}
						}
					}
				}
			}
		},
...
		"/alarms/{alarmId}": {
			"get": {
				"operationId": "GetAlarm",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/IAlarm"
										}
									],
									"nullable": true
								}
							}
						}
					},
					"404": {
						"description": "the requested Alarm in not found"
					}
				},
				"description": "Get a Alarm details",
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "alarmId",
						"required": true,
						"schema": {
							"type": "string"
						},
						"example": "63133c886978820e456b8015"
					}
				]
			},
			"delete": {
				"operationId": "DeleteAlarm",
				"responses": {
					"200": {
						"description": "Deleted",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/Document_any.any.any_"
										},
										{
											"$ref": "#/components/schemas/IAlarm"
										},
										{
											"$ref": "#/components/schemas/Required_any_"
										}
									]
								}
							}
						}
					},
					"404": {
						"description": "the requested alarm in not found"
					}
				},
				"description": "Delete a alarm",
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "alarmId",
						"required": true,
						"schema": {
							"type": "string"
						},
						"example": "63133c886978820e456b8015"
					}
				]
			}
		},
		"/alarms/create": {
			"post": {
				"operationId": "CreateAlarm",
				"responses": {
					"200": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IAlarm"
								},
								"examples": {
									"Example 1": {
										"value": {
											"name": "new therapist",
											"referenceType": "Agreement",
											"referenceID": "63131c361ea32f520cb28bd6",
											"frequencyUnit": "Days",
											"frequency": 15,
											"active": true,
											"startDate": "2022-09-10T00:00:00.000Z",
											"endDate": "2022-09-10T00:00:00.000Z"
										}
									}
								}
							}
						}
					},
					"422": {
						"description": "Validation Failed"
					}
				},
				"description": "Create a alarm",
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IAlarm"
							}
						}
					}
				}
			}
		},
		"/alarms/update/{alarmId}": {
			"put": {
				"operationId": "UpdateAlarm",
				"responses": {
					"200": {
						"description": "updated",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/IAlarm"
										}
									],
									"nullable": true
								}
							}
						}
					},
					"422": {
						"description": "Validation Failed"
					}
				},
				"description": "Update a alarm",
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "alarmId",
						"required": true,
						"schema": {
							"type": "string"
						},
						"example": "63133c886978820e456b8015"
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Partial_IAlarm_"
							},
							"example": "63133c886978820e456b8015"
						}
					}
				}
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}